Fix `sub_materialize` for GPU arrays (#261)
Currently, `sub_materialize` (through `sub_materialize_axes`) falls back
to materializing on CPU. This PR generalizes that logic by determining
the output destination with `similar`, which helps to support non-Array
types like GPU arrays. As a stand-in for other GPU arrays, I test this
using
[JLArrays.JLArray](https://github.com/JuliaGPU/GPUArrays.jl/tree/master/lib/JLArrays),
which is a reference implementation for the GPUArrays.jl interface that
runs on CPU.

An alternative design would be to define memory layouts for GPU arrays
(i.e. #9), which would allow more customizability for GPU array
backends, however I think it is helpful to have fallbacks that "just
work" if reasonable parts of the Base AbstractArray interface are
implemented.

I hit this issue because I was testing out
[`BlockArrays.BlockedArray`](https://juliaarrays.github.io/BlockArrays.jl/stable/man/blockedarrays)
wrapping a GPU array and noticed that calling `A[Block(1, 1)]` to access
a block instantiated the block on CPU, this PR fixes that issue.

---------

Co-authored-by: Sheehan Olver <solver@mac.com>